requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
configparser==6.0.0
dnspython==2.4.1
docutils==0.20.1
idna==3.4
lockfile==0.12.2
pid==3.0.4
pip==22.3.1
python-daemon==3.0.1
requests==2.31.0
service==0.6.0
setproctitle==1.3.2
setuptools==65.5.1
urllib3==2.0.4
wheel==0.38.4
